Refactor persisted state into one bound store
