+++ src/components/Settings/Settings.js
@@ -22,7 +22,7 @@ import {
   useTranslation,
   useUserMe
 } from "sharedHooks";
-import usePersistedStore from "stores/usePersistedStore.ts";
+import useStore from "stores/useStore";
 
 const SETTINGS_URL = `${Config.OAUTH_API_URL}/users/edit?noh1=true`;
 const FINISHED_WEB_SETTINGS = "finished-web-settings";
@@ -32,8 +32,8 @@ const Settings = ( ) => {
   const { t } = useTranslation();
   const currentUser = useCurrentUser( );
   const { remoteUser, isLoading, refetchUserMe } = useUserMe();
-  const isAdvancedUser = usePersistedStore( state => state.isAdvancedUser );
-  const setIsAdvancedUser = usePersistedStore( state => state.setIsAdvancedUser );
+  const isAdvancedUser = useStore( state => state.isAdvancedUser );
+  const setIsAdvancedUser = useStore( state => state.setIsAdvancedUser );
 
   const [settings, setSettings] = useState( {} );
   const [isSaving, setIsSaving] = useState( false );
