+++ src/stores/createLayoutSlice.js
@@ -1,0 +1,6 @@
+const createLayoutSlice = set => ( {
+  isAdvancedUser: false,
+  setIsAdvancedUser: newValue => set( { isAdvancedUser: newValue } )
+} );
+
+export default createLayoutSlice;
