+++ tests/integration/navigation/ARCamera.test.js
@@ -5,7 +5,7 @@ import {
 } from "@testing-library/react-native";
 import * as usePredictions from "components/Camera/ARCamera/hooks/usePredictions";
 import initI18next from "i18n/initI18next";
-import usePersistedStore from "stores/usePersistedStore.ts";
+import useStore from "stores/useStore";
 import { renderApp } from "tests/helpers/render";
 import setupUniqueRealm from "tests/helpers/uniqueRealm";
 
@@ -52,7 +52,7 @@ afterAll( uniqueRealmAfterAll );
 beforeAll( async () => {
   await initI18next();
   jest.useFakeTimers( );
-  usePersistedStore.setState( { isAdvancedUser: true } );
+  useStore.setState( { isAdvancedUser: true } );
 } );
 
 describe( "ARCamera navigation with advanced user layout", ( ) => {
