+++ tests/integration/navigation/SoundRecorder.test.js
@@ -4,7 +4,7 @@ import {
   within
 } from "@testing-library/react-native";
 import initI18next from "i18n/initI18next";
-import usePersistedStore from "stores/usePersistedStore.ts";
+import useStore from "stores/useStore";
 import { renderApp } from "tests/helpers/render";
 import setupUniqueRealm from "tests/helpers/uniqueRealm";
 
@@ -36,7 +36,7 @@ afterAll( uniqueRealmAfterAll );
 beforeAll( async () => {
   await initI18next();
   jest.useFakeTimers( );
-  usePersistedStore.setState( { isAdvancedUser: true } );
+  useStore.setState( { isAdvancedUser: true } );
 } );
 
 describe( "SoundRecorder navigation", ( ) => {
