+++ tests/integration/navigation/StandardCamera.test.js
@@ -5,7 +5,7 @@ import {
   within
 } from "@testing-library/react-native";
 import initI18next from "i18n/initI18next";
-import usePersistedStore from "stores/usePersistedStore.ts";
+import useStore from "stores/useStore";
 import { renderApp } from "tests/helpers/render";
 import setupUniqueRealm from "tests/helpers/uniqueRealm";
 
@@ -37,7 +37,7 @@ afterAll( uniqueRealmAfterAll );
 beforeAll( async () => {
   await initI18next();
   jest.useFakeTimers( );
-  usePersistedStore.setState( { isAdvancedUser: true } );
+  useStore.setState( { isAdvancedUser: true } );
 } );
 
 describe( "StandardCamera navigation", ( ) => {
